Service1/requirements.txt
boto3==1.35.33
botocore==1.35.33
Flask==3.0.3
flask_sqlalchemy==3.1.1
pika==1.3.2
python-dotenv==1.0.1
SQLAlchemy==2.0.35
PyMySQL
